pyproject.toml
[project]
name = "freelance"
version = "0.1.0"
description = ""
authors = [
    {name = "aLMASIK",email = "suleimenovalmas75@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
